Testing pawn
Stared testing the differnet pawns in vr to check if they are working correctly. More or less finished the idea of the measurent pawn. Is a problem with the rotation gain implementation. Translation works fine. Need to calculate gain values for curvature. Staging stage works just need to add some additionnal feature to make it better.