Created function for alligning virtual and physical space.
The physical and virtual space can now be aligned when using translation and curvature gain. Further improved the code for handling movement back and forth in the threhsold experiment. Need to add a better target class to trigger when the player has moved 5m in world space in the virtural world.